% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addText.R
\name{addText}
\alias{addText}
\alias{text2}
\title{Add text to an existing lattice plot}
\usage{
addText(x, y, labels, col = "red", cex = 2/3, adj = 0.5, ...)
}
\arguments{
\item{x, y}{Numeric vectors of coordinates where text \code{labels} should
be written. If \code{y} is missing, \code{x} is treated as a list with 'x' and 'y'}

\item{labels}{Character vector or expression specifying the text to
be written. If \code{x} is specified as a list and \code{y} is a character
vectors, \code{labels} will use the values in \code{y}}

\item{col, cex, adj}{Default values passed to \code{\link[lattice:panel.text]{lattice::panel.text()}}}

\item{...}{Other parameters are passed to \code{\link[lattice:panel.text]{lattice::panel.text()}}}
}
\value{
Nothing is returned. This function is called for the side effect of
adding labels.
}
\description{
This serves the same role as the base function \code{\link[graphics:text]{graphics::text()}} for
use in \code{lattice} plots.  (Also can be called as \code{text2()}.)
}
\details{
Text labels can be added to an existing simple lattice plot with this
function. Values in \code{x}, \code{y} and \code{labels} are recycled as needed.
Interaction with the existing plot occurs through the function
\code{\link[lattice:trellis.currentLayout]{lattice::trellis.currentLayout()}}. See that function for more details.
}
\examples{

}
