% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFres.R
\name{getFres}
\alias{getFres}
\alias{get.p}
\title{Extract a Filter Result}
\usage{
getFres(
  x,
  filter,
  what = c("p", "true", "count", "n", "positive", "negative"),
  as.percent = FALSE,
  excluded = FALSE,
  fmt = "\%4.1f\%\%"
)

get.p(x, filter, as.percent = FALSE, excluded = FALSE)
}
\arguments{
\item{x}{Either a \code{filterResult} or a \code{flowFrame}
or a \code{flowSet}}

\item{filter}{An optional \code{filter}, required if \code{x} is not a
\code{filterResult}}

\item{what}{A single character string specifying the statistic to
return. The default value of \code{"p"} extracts the positive
fraction. (See the Details section for other options)}

\item{as.percent}{If \code{TRUE}, return the positive fraction (or
negative fraction) as a formatted character string}

\item{excluded}{If \code{TRUE}, return the negative fraction or
negative population}

\item{fmt}{Character format string for \code{\link{sprintf}} if the
argument \code{as.percent = TRUE}}
}
\value{
For \code{getFres}: a numeric vector of the desired statistic \emph{or}
a character string formatted as the percent of positive or negative events.

For \code{get.p}: a numeric vector of the positive fraction, \emph{or} 
a character string formatted as the percent positive events.
}
\description{
Extract a single summary statistic from a \code{filterResult} or
\code{flowFrame/flowSet} plus \code{filter}
}
\details{
This function will extract a selected statistic from a filter
result. If needed, it will apply a \code{filter} to a
\code{flowFrame} or \code{flowSet} and then extract the desired
statistic. The positive or negative population can be specified.


This is a convenience function to extract information normally produced
by sequentially calling \code{filter}, \code{summary} and an extraction
function to retrieve slot values. The extracted statistic is either the
positive fraction (\code{p}), positive number (\code{true}) or
total number (\code{count}).

If the first argument (\code{x}) is a \code{filterResult} or \code{
filterResultList}, the argument \code{filter} is ignored \emph{unless}
it is a single character string that will be interpreted as \code{what}.
This allows for lazy applications of the function to filter results
such as \code{getFres(res, "n")} to extract the total number of
events from the object \code{res}. If the first argument is 
a \code{flowFrame} or \code{flowSet}, \code{filter} must be
provided in order to apply it to the data before extracting
the statistic.

The specific statistic can be specified by the character value in
\code{what} as indicated below:
\tabular{ll}{
	\code{"p"} \tab The positive fraction of events. \cr
	\code{"true"} \tab Number of positive events. \cr
	\code{"positive"} \tab Number of positive events. \cr
	\code{"negative"} \tab Number of negative events. \cr
	\code{"n"} \tab Total number of events. \cr
	\code{"count"} \tab Total number of events. \cr
}
Note that with \code{what = "n"} or \code{what = "count"},
the \code{excluded} option is ignored.
}
