% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGate.R
\name{linearGate}
\alias{linearGate}
\title{Fit Gate for Single Cells}
\usage{
linearGate(x, xchan = "FL2.A", ychan = "FL2.H",
  zero.intercept = TRUE, dx = 0.06, dy = dx, xrange = NULL,
  yrange = NULL, cutoff = 0.02, limit.gate = FALSE,
  filterId = "singlets", groupFilterId = "linearGateList")
}
\arguments{
\item{x}{Data to fit as either a matrix from \code{\link[Biobase]{exprs}},
a \code{flowFrame}, or a \code{flowSet}}

\item{xchan, ychan}{Character strings to define single cells, default
values of \code{"FL2.A"} and \code{"FL2.H"}}

\item{zero.intercept}{If \code{TRUE} (default), force regression through
origin}

\item{dx, dy}{Width and height of rectangular gate, default value of 0.06.
If only \code{dx} is specified, \code{dy} will be assigned the same value}

\item{xrange, yrange}{If \code{NULL}, regression will be limited to
values within the 2.5 - 97.5 percentile, otherwise in the specified
range of values}

\item{cutoff}{Numeric value between 0 and 1 specifying the fraction of
low and high values to ignore in the regression, default value of 0.02}

\item{limit.gate}{Logical value (default of \code{FALSE}) to limit the
returned gate by the values in \code{xrange} and \code{yrange}}

\item{filterId}{Character string with default value of \code{"singlets"}}

\item{groupFilterId}{Character string with default value of \code{
"linearGateList"}}
}
\value{
A rectangular \code{polygonGate} or list of these gates.
}
\description{
Fit a rectangular polygon gate for single cells in FL2.A x FL2.H
}
\details{
This function fits values in \code{xchan} and \code{ychan} to a robust
linear regression with \code{\link[MASS]{lqs}}. Note that default names
for the x & y channels are "FL2.A" and "FL2.H" and \strong{not} "FL2-A"
and "FL2-H". If \code{zero.intercept} is \code{TRUE}, the fit is forced
through origin. The regression will be limited to values within the
2.5 to 97.5 percentile if \code{xrange} and \code{yrange} is \code{NULL}.
A different range can be specified with these arguments.

The argument \code{x} can be a matrix from \code{\link[Biobase]{exprs}},
a \code{flowFrame}, or a \code{flowSet}. A recantgular polygon gate (or
list of gates) with a width and height of \code{dx} and \code{dy} will
be returned. If \code{limit.gate} is \code{TRUE}, the returned gate will
also be limited by the values in \code{xrange} and \code{yrange}.
Otherwise, the gate will be limited by the range of values in each
flowFrame up to limits set by \code{cutoff} and the maximum range of
values in the x channel (FL2.A) limited to \code{cutoff * maxRange} and
\code{(1 - cutoff) * maxRange}.
}
