% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locator2.R
\name{locator2}
\alias{locator2}
\title{Locator Function for Lattice}
\usage{
locator2(n = 512, type = c("p", "v", "h", "both", "none"), ...)
}
\arguments{
\item{n}{The maximum number of points to locate}

\item{type}{One of \code{"p", "v", "h", "both"} or 
\code{"none"} to add a point, vertical line, horizontal line,
both vertical and horizontal lines, or no mark at each selected
point}

\item{...}{Additional graphic parameters passed to either
\code{\link{panel.points}} or \code{\link{panel.abline}} if
\code{type != "none"}}
}
\value{
A list containing \code{x} and \code{y} coordinates of the
identified points in the user coordinate system.
}
\description{
Reads the position of the graphics cursor from a lattice panel
}
\details{
This provides a functional replacement for
\code{\link{locator}} that works with \code{lattice} objects. A
\code{lattice} panel will be selected, redrawn as a single panel and
used to interact with the user. The position of the graphics cursor
will be recorded with each click of the (first) mouse button.

By default, a point will be added at that position unless otherwise
specified by the argument \code{type}.

Like the \code{\link{locator}} function, this function returns 
values in \code{"usr"} coordinates as a list.
}
