% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFit.R
\name{peakFit}
\alias{peakFit}
\alias{mgfun}
\alias{gfun}
\title{Fit G1 and G2 peaks}
\usage{
peakFit(x)

mgfun(x, mean, sd = 0.05 * mean, lambda = 1/length(mean))

gfun(x, mean, sd)
}
\arguments{
\item{x}{Numeric values at which to evaluate the function}

\item{mean}{Mean value for the Gaussian distribution}

\item{sd}{Numeric value of standard deviation (sigma) for Gaussian
function}

\item{lambda}{Numeric vector serving as the multiplier for
the relative area of each Gaussian component with a default value of 
`1/length(mean)` for each. The value will be adjusted to
`lambda/sum(lambda)` to permit a value such as `c(2, 5)` to express the
relative **area** of each peak.}
}
\value{
Echo argument

The combined Gaussian density for given `x` values and parameters

The Gaussian density for given `x` values
}
\description{
Fit an arbitrary number of Gaussian peaks.

Density distribution function for multiple Gaussian function of 
given means (`mean`) and standard deviations (`sd`), each adjusted  by
the multiplier `lambda` where `sum(lambda) = 1`.

This is a simpler version of [stats::dnorm] and runs a little faster.
}
\details{
Currently a stub for the \code{peakFit} function.
}
