% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSet.R
\name{readSet}
\alias{readSet}
\title{Read and Transform Flow Set}
\usage{
readSet(
  path = ".",
  pattern = "\\\\d\\\\d\\\\d$",
  log.chan = NULL,
  alter.names = TRUE,
  name.keyword = "SAMPLE ID",
  phenoData = list(date = "$DATE", sample = "SAMPLE ID"),
  tfun = c("asinh", "log10"),
  depth = 10,
  log.cutoff = 2^depth,
  ...
)
}
\arguments{
\item{path}{A character vector identifying the path to search for FSC
files, defaults to current directory}

\item{pattern}{Grep pattern to identify FSC files, defaults to
\code{"\\d\\d\\d$"}}

\item{log.chan}{Character vector of channels to be log-transformed (after
the \code{"alter.names"} option is applied)}

\item{alter.names}{Logical value passed to \code{read.flowSet} to 
change "-" to "." in names, default value of \code{TRUE}}

\item{name.keyword}{Character string for the  FCS keyword used for names,
default of \code{"SAMPLE ID"}}

\item{phenoData}{List to serve as the rudiments of the \code{phenoData} 
built from the keywords \code{"$DATE"} and \code{"SAMPLE ID"}}

\item{tfun}{Transformation function to apply to log-transformed values,
either \code{asinh} (default) or \code{log10}}

\item{depth}{Integer indicating bit depth (\code{2^depth}), 10 for
FACSCaliber, 24 for Acurri}

\item{log.cutoff}{Values greater than this cutoff force transformation if 
the argument \code{log.chan} is \code{NULL}, defaults to \code{2^depth}}

\item{...}{Additional arguments to \code{\link[flowCore]{read.flowSet}}}
}
\value{
Flow Set with transformed and name-adjusted parameters.
}
\description{
A wrapper to \code{\link{read.flowSet}} to read a flow set.
}
\details{
Files in the directory specified by \code{path} that match 
the \code{grep} pattern specified in \code{pattern} will be read with
\code{\link{read.flowSet}}. Parameter names will be adjusted if
\code{alter.names = TRUE} to replace \code{"-"} with \code{"."}. A 
minimal amount of \code{phenoData} will be assigned from the $DATE
and "SAMPLE ID" keywords.
}
