% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaplot.R
\name{dnaplot}
\alias{dnaplot}
\title{Densityplot for DNA Profiles}
\usage{
dnaplot(
  ...,
  darg = list(bw = "SJ", n = 512, adjust = 1),
  chan = NULL,
  plot = TRUE
)
}
\arguments{
\item{...}{Arguments to be handed to \code{\link[flowViz:densityplot]{flowViz::densityplot()}} \emph{or}
a \code{flowFrame} or \code{flowSet}}

\item{darg}{List of arguments passed to the kernel density function
\code{\link[stats:density]{stats::density()}} to increase the resolution of the plot. In
addition to \code{bw}, these can include \code{adjust}, \code{n}. The option
\code{na.rm = TRUE} will be included in all cases}

\item{chan}{An option character scalar identifying the data to be
plotted if \code{...} is a \code{flowFrame} or \code{flowSet}}

\item{plot}{If \code{TRUE}, lattice object will be plotted as well as
being returned as an \code{invisible} object}
}
\value{
Optionally plot the data as a side-effect and invisibly
return the lattice plot.
}
\description{
A wrapper for \code{\link[flowViz:densityplot]{flowViz::densityplot()}} with defaults for
DNA profiles.
}
\details{
This is a convenience function with parameters more suited for
typical DNA profiles by using the bandwidth algorithm of Sheather
& Jones in \code{bw.SJ} with the adjustment in \code{adjust} and using more
points (512 rather than 50) to determine the density. These additional
parameters are passed to \code{\link[stats:density]{stats::density()}}, which in \code{darg}.

The call effectively executes the following:\preformatted{  densityplot(..., dargs = dargs)
}

... \strong{except} for the \emph{lazy} case of calling it with first argument a
\code{flowSet} or \code{flowFrame}. In this case the function guesses
the most likely \code{chan} parameter (in order) from \verb{FL2.A, FL3.A, FL2.H}
or \code{FL3.H}. Alternatively, the channel name can be passed in \code{chan} and
the call now becomes:\preformatted{  densityplot(~ chan, ..., dargs = dargs)
}
}
\examples{
# load and filter example data with 16 flowFrames
  fs <- readSet(system.file("extdata", "synch/", package = "flowExtra"))
  fs <- Subset(fs, boundaryFilter("FL2.A"))
# default density parameters from flowViz 
  dnaplot(~ FL2.A, fs, xlim = c(50, 500), darg = NULL,
    main = "Default parameters for densityplot")
# revised density parameters 
  dnaplot(~ FL2.A, fs, xlim = c(50, 500),
    main = "Adjusted parameters for 'dnaplot'")
# lazy use if "FL2.A" is present
  dnaplot(fs)

}
