% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaplot.R
\name{dnaplot}
\alias{dnaplot}
\title{Densityplot for DNA Profiles}
\usage{
dnaplot(..., adjust = 0.2, n = 512, plot = TRUE)
}
\arguments{
\item{...}{Arguments to be handed to \code{\link[flowViz:densityplot]{flowViz::densityplot()}}}

\item{adjust}{Numeric factor to adjust the bandwith in the kernel
density function \code{\link[stats:density]{stats::density()}}, decreased from 1 to 0.2}

\item{n}{The number of equally spaced points at which the density is to
be estimated, increased from 50 to 512}

\item{plot}{If \code{TRUE}, lattice object will be plotted as well as
being returned as an \code{invisible} object}

\item{data}{\code{flowFrame} or \code{flowSet} with data}
}
\value{
Optionally plot the data as a side-effect and invisibly
return the lattice plot.
}
\description{
A wrapper for \code{\link[flowViz:densityplot]{flowViz::densityplot()}} with defaults for
DNA profiles.
}
\details{
This is a convenience function with parameters more suited for
typical DNA profiles. Specifically, the optional arguments to
\code{\link[stats:density]{stats::density()}}, which are provided in \code{darg}, are changed from
\code{n = 50} and \code{adjust = 1} to \code{n = 512} and \code{adjust = 0.2}.

The call effectively executes the following:\preformatted{  densityplot(..., dargs = list(adjust = adjust, n = n))
}
}
\examples{
 fs <- readSet(system.file("extdata", "RPE_synch/", package = "flowExtra"))
 dnaplot(~ FL2.A, fs) # unadjusted values

}
