% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFres.R
\name{getFres}
\alias{getFres}
\alias{get.p}
\title{Extract a Filter Result}
\usage{
getFres(x, filter, what = c("p", "true", "count", "n", "positive",
  "negative"), as.percent = FALSE, excluded = FALSE,
  fmt = "\%4.1f\%\%")

get.p(x, filter, as.percent = FALSE, excluded = FALSE)
}
\arguments{
\item{x}{Either a \code{filterResult} \emph{or} a \code{flowFrame}
or \code{flowSet}}

\item{filter}{An optional \code{filter}, required if \code{x} is not a
\code{filterResult}}

\item{what}{A single character string specifying the statistic to
return. The default value of "p" extracts the positive fraction.
(see Details for other options)}

\item{as.percent}{If \code{TRUE}, return the positive fraction (or
negative fraction) as a formatted character string}

\item{excluded}{If \code{TRUE}, return the negative fraction or
negative population}

\item{fmt}{Character format string for \code{sprintf} if the argument
\code{as.percent = TRUE}}
}
\value{
For \code{getFres}: a numeric vector of the desired statistic \emph{or}
a character string formatted as the percent of positive or negative events.

For \code{get.p}: a numeric vector of the positive fraction, \emph{or} 
a character string formatted as the percent positive events.
}
\description{
Extract a single summary statistic from a \code{filterResult} or
\code{flowFrame/flowSet} plus \code{filter}
}
\details{
This function will extract a selected statistic from a filter
result. If needed, it will apply a \code{filter} to a
\code{flowFrame} or \code{flowSet} and then extract the desired
statistic. The positive or negative population can be specified.


This is a convenience function to extract information normally produced
by sequentially calling \code{filter}, \code{summary} and an extraction
function to retrieve slot values. The extracted statistic is either the
positive fraction (\code{p}), positive number (\code{true}) or
total number (\code{count}).

If the first argument (\code{x}) is a \code{filterResult} or \code{
filterResultList}, the argument \code{filter} is ignored \emph{unless}
it is a single character string that will be interpreted as \code{what}.
This allows for lazy applications of the function to filter results
such as \code{getFres(res, "n")} to extract the number of positive
events from the object \code{res}. If the first argument is 
a \code{flowFrame} or \code{flowSet}, \code{filter} must be
provided in order to apply it to the data before extracting
the statistic.

The specific statistic can be specified by several methods. The default
value is "p" for the positive fraction. The negative fraction can be
extracted with the option \code{excluded = TRUE}. If \code{as.percent =
TRUE}, the value will be returned as a character string formatted with
\code{\link{sprintf}} according the string in \code{fmt}. The number of
positive events can be retrieved by \code{what = "true"} or \code{what =
"positive"}. The total number of events can be retrieved with
\code{what = "count"}. The number of negative events can be retrieved by
\code{what = "negative"}. Note that with these last two options, 
the \code{excluded} option is ignored.
}
