% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearGate.R
\name{linearGate}
\alias{linearGate}
\title{Create Gate for Single Cells}
\usage{
linearGate(
  x,
  xchan = "FL2.A",
  ychan = "FL2.H",
  zero.intercept = TRUE,
  width = 0.05,
  xRange = NULL,
  gRange = TRUE,
  filterId = "singlets",
  groupFilterId = "singletsGateList"
)
}
\arguments{
\item{x}{Either a matrix from \code{\link[Biobase]{exprs}}, a
\code{flowFrame}, or a \code{flowSet}}

\item{xchan, ychan}{Character strings identifying the data to search for
single cells with default values of \code{"FL2.A"} and \code{"FL2.H"}}

\item{zero.intercept}{If \code{TRUE} (the default), the regression used
to identify the singlets will be forced through the origin}

\item{width}{The width of the rotated rectangular gate as a fraction
of the instrument data range for the \code{xchan} parameter,
with a default value of 5 \% (0.05)}

\item{xRange}{A numeric value of length two to specify the upper and
lower percentile range of the data in \code{xchan} to be used for the
regression. If \code{NULL}, a default value of \code{c(0.025, 0.975)}
will be used}

\item{gRange}{A \code{logical} value \emph{or} a numeric vector of length
2 to specify limits on the dimension of the returned gate. The limits
can be specified by quantiles or absolute values. See the Details
section for more information}

\item{filterId}{Character string with default value of \code{"singlets"}}

\item{groupFilterId}{Character string with default value of
\code{"singletsGateList"}}
}
\value{
A rectangular \code{polygonGate} or list of these gates.
}
\description{
Fit a narrow rectangular gate to select single cells in a 2-D
plot of area x height
}
\details{
This function creates a data-driven gate to select single cells from
\code{area} and \code{height} values collected on a linear scale. A
robust linear regression is fit to the data with
\code{\link[MASS]{lqs}}. This regression is used to create a narrow
rectangular gate centered about the single cells. If
\code{zero.intercept = TRUE}, the regression is forced through origin.
By default, the regression is applied to values within the 2nd to 98th
percentile of data in channel \code{xchan}. Other percentiles can be
specified with the argument \code{xRange}.

The resulting \code{polygonGate(s)} will be limited to the range of
values specified by \code{gRange}. If \code{gRange = TRUE}, the gate
will be limited to the 2.5 to 97.5 percentile of data in
channel \code{xchan}. If \code{gRange = FALSE}, the gate will
span the full range of data. If \code{gRange} is a numeric vector of
values less than or equal to 1, \code{gRange} will be treated as
the percentiles to limit the gate. If \code{gRange} is a numeric
vector with values greater than 1, \code{gRange} will be
interpreted as the lower and upper limits of the gate in the \code{xchan}
dimension.

Since this is typically used for cell cycle analysis with DNA
content, the default parameters are \code{xchan = "FL2.A"} and
\code{ychan = "FL2.H"}
}
