% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFind.R
\name{peakFind}
\alias{peakFind}
\title{Find G1 and G2 peaks}
\usage{
peakFind(x, curveFilter = NULL, chan = "FL2.A", searchFun = NULL,
  range.search = NULL, probs = c(0.05, 0.95), ...)
}
\arguments{
\item{x}{A \code{flowFrame} or \code{flowSet} with the parameter named in \code{chan}}

\item{curveFilter}{Optional function that returns a \code{multipleFilterResult}
defining the regions of 1-D curvature in \code{chan}. If \code{NULL}, the
\code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}} will be used}

\item{chan}{Name of the data to be evaluated in \code{x}, default of \code{"FL2.A"}}

\item{searchFun}{Optional function to determine the position of each
peak in defined by the argument \code{curvFilter}. If \code{NULL}, the
\code{\link[stats:density]{stats::density()}} function will be used with the bandwith adjusted
by the option \code{adj = 2}}

\item{range.search}{Optional numeric vector of length 2 defining the
lower and upper limits to accept peaks. The default value of \code{NULL}
limits the search to the range defined by the quantiles specified
in \code{probs}}

\item{probs}{Numeric vector of length 2 defining the lower and upper
quantiles for valid peak values, the default value of 5 to 95\%
is specified by the argument \code{probs = c(0.05, 0.95)}}

\item{...}{Additional arguments to be passed to
\code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}}}
}
\value{
An 1-D or 2-D array with rownames obtained from \code{\link[flowCore:identifier]{flowCore::identifier()}}
and column names determined as described above.
}
\description{
Find G1 and G2 (and other) peaks in FL2.A values
}
\details{
This function identifies the probable location of the G1 and G2/M peaks
in the \code{"FL2.A"} channel (or the channel
specified by the argument \code{chan}) in a \code{flowFrame} or \code{flowSet}. Areas
of high density are identified by the function defined in \code{curveFilter}. If
this is \code{NULL}, the \code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}} function will be used
with optional arguments (\code{bwFac} or \code{gridsize}) passed in \code{...}.

By default, peaks are accepted in the range defined by the 5th to 95th
percentile or to the percentile values specified in \code{probs}. For
well-behaved data, this can be set \code{c(50, 500)} to exclude apoptotic
populations and populations greater than G2/M. If no more than two peaks
are found, they will be labeled \code{"G1"} and \code{"G2"}. A single peak will be
labeled as \code{"G1"} with an \code{NA} for the second peak, labeled as \code{"<G2>"}.
If more than two peaks are found, all peaks will be labeled as
\code{"peak1"}, \code{"peak2"}, \code{"peak3"}, etc.
}
