% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFind.R
\name{peakFind}
\alias{peakFind}
\title{Find G1 and G2/M peaks}
\usage{
peakFind(
  x,
  chan = "FL2.A",
  darg = list(bw = "nrd0", n = 512),
  range.search = NULL,
  probs = c(0.05, 0.95),
  bwFac = 2,
  gridsize = rep(401, 2)
)
}
\arguments{
\item{x}{A \code{flowFrame} or \code{flowSet} with the parameter named in \code{chan}
\emph{or} a \code{numeric} vector or matrix (see Details)}

\item{chan}{The name of the data to be evaluated in \code{x}}

\item{darg}{List of arguments passed to the kernel density estimating
function (\code{\link[stats:density]{stats::density()}}) used to find peaks}

\item{range.search}{An optional numeric vector of length 2 setting the
range of values in \code{chan} to search for peaks. The default value of \code{NULL}
limits the search to the range defined by the quantiles specified
in \code{probs}}

\item{probs}{Numeric vector of length 2 defining the lower and upper
quantiles to set the value for \code{range.search} if not provided}

\item{bwFac, gridsize}{Numeric values of length 1 and 2 used by
\code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}} to identify regions of 1-D curvature in
the data (\code{bwFac} increased from default of 1.2 to 2)}
}
\value{
An 2-D array with row names obtained from \code{\link[flowCore:identifier]{flowCore::identifier()}}
and column names determined as described above.
}
\description{
Find G1 and G2/M (and other) peaks in linear values
}
\details{
This function identifies the probable location of the G1 and G2/M peaks
in the channel identified by \code{chan} in either a \code{flowFrame} or \code{flowSet}
or "raw" values as described below.
Areas of local high density are first identified by
\code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}} with arguments in \code{bwFac} and \code{gridsize}.
The values are the default values for by \code{curv1Filter()}.
The peak location within each local area of high density is
identified from a kernel density estimate \code{\link[stats:density]{stats::density()}} using
arguments passed in the list \code{darg}. Peak positions are returned as
an array of two or more columns.

If \code{range.search} is \code{NULL}, peaks are accepted in the range of data
spanning 5th to 95th percentile or as specified by the percentile values
in \code{probs}. For well-behaved data, \code{range.search} can be set
to \code{c(50, 500)} to exclude apoptotic and polyploid populations.

If exactly two peaks are found, they will be labeled \code{"G1"} and \code{"G2"}.
If a mixture of one and two peaks are found, the single peak will be
treated as the \code{"G1"} peak and the \code{"G2"} peak will be assigned an \code{NA}
value. If only one peak is found or if more than two peaks
are found, the peaks will be labeled as \code{"peak1"}, \code{"peak2"}, \code{"peak3"},
etc. A warning message will report finding other than 2 peaks.

The argument \code{x} can also be a \code{matrix} or \code{numeric} vector of
"raw" values which will be
coerced into a matrix. If the name of a column (or vector) is \code{chan},
that parameter will be analyzed. If the \code{matrix} or \code{vector} is not
named, the first column will be analyzed.
}
\examples{
 fs <- readSet(system.file("extdata", "synch/", package = "flowExtra"))
 peakFind(fs) # find all by default
 peakFind(fs[9:16]) # paired peaks are labeled "G1" and "G2"

}
