% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakNormalize.R
\name{peakNormalize}
\alias{peakNormalize}
\title{Translate and Scale G1 and G2 Peaks}
\usage{
peakNormalize(fs, peaks = NULL, g1 = 200, g2 = g1 * ratio,
  ratio = 1.9, chan = "FL2.A", limits = c(0, 1023),
  range.search = c(50, 500), only.G1 = FALSE, ...)
}
\arguments{
\item{fs}{A \code{flowSet} with data to be transfomred in channel in \code{chan}}

\item{peaks}{An optional matrix of "G1" and "G2" peak values \emph{or} a
vector "G1" values. The default value of \code{NULL} will use
\code{\link[=peakFind]{peakFind()}} to identify the peaks}

\item{g1}{Value of the rescaled G1 peak (\code{200})}

\item{g2}{Value of the rescaled G2 peak (\code{ratio * g1})}

\item{ratio}{G2 to G1 ratio, used to scale data and place missing peaks
(\code{1.90})}

\item{chan}{Character vector to specify the channel to normalize
(\code{"FL2.A"})}

\item{limits}{Numeric vector of length 2 specifying the limits for the
transformed data (default value of \code{c(0, 1023)}). If \code{NULL}, the range
will the the original limits of the data. If \code{FALSE} the range will be
\code{-Inf} to \code{+Inf}.}

\item{range.search}{A numeric vector of length 2 specifying the range
of values in the orignal data in which peaks will be accepted. The
default value of \code{c(50, 500)} is chosen to ignore apoptotic values and
polyploid peaks}

\item{only.G1}{\code{Logical} value to prevent scaling by only using the
"G1" peak to anchor the data}

\item{...}{Additional arguments passed to \code{\link[=peakFind]{peakFind()}} and then to
\code{\link[flowStats:curv1Filter]{flowStats::curv1Filter()}}}
}
\value{
A transformed and (optionally) scaled \code{flowSet}.
}
\description{
Normalize "G1" and "G2" peaks in DNA profile data
}
\details{
This function attempts to translate a linear DNA profile data to align
peaks corresponding to a "G1" population of cells. The "G2" peak will
be adjusted by scaling to be positioned at the value given by \code{ratio * g1} where \code{g1} is the position of the "G1" peak. Although this ratio
should theoretically be 2.0, it is empirically closer to 1.9.

The position of the actual peaks is specified in the optional argument
\code{peaks}. If provided, this can either be a vector of "G1" peak
positions or a matrix of "G1" and "G2" peak positions. A larger matrix
can be accepted, such as one produced with \code{\link[=peakFind]{peakFind()}}, but only the
first two columns will be used with "G1" and "G2" in the first and second
columns, respectively. Missing values in either "G1" or "G2"
will be replaced by a value derived from \code{ratio}. If the argument \code{ peaks} is \code{NULL}, the function \code{\link[=peakFind]{peakFind()}} will be called with
arguments \code{range.search} and \code{...} to identify the peaks.

If the \code{logical} argument \code{only.G1 = TRUE}, a complete set of "G1"
values must be provided and the data will only be translated to place
the "G1" peak at the selected value.

The transformed values will be trimmed to the range specified by \code{ limits}, which has a default value of \code{c(0, 1023)}. The full range of
original values will be used if \code{limits = NULL} and no trimming will
occur if \code{limits = FALSE}.
}
