% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeGate2.R
\name{rangeGate2}
\alias{rangeGate2}
\title{Improved search for separation between positive and negative populations}
\usage{
rangeGate2(
  fs,
  stain,
  from = NULL,
  to = NULL,
  cutoff = 0.05,
  adjust = 1,
  half.range = TRUE,
  sd = 2.5,
  positive = TRUE,
  method = c("minimum", "left"),
  plot = FALSE,
  legend = TRUE,
  filterId = "defaultRectangleGate",
  ...
)
}
\arguments{
\item{fs}{A \code{\link[flowCore:flowset]{flowCore::flowset}} or \code{\link[flowCore:flowFrame]{flowCore::flowFrame}} object}

\item{stain}{A \code{character} scalar (also known as \code{chan}) of the flow
parameter to analyze}

\item{from, to}{Starting and ending values to search for the break}

\item{cutoff}{Peaks and valleys smaller than this fraction of the range
will be ignored, default value of 0.05}

\item{adjust}{Adjustment to bandwidth used by \code{\link[stats:density]{stats::density()}}}

\item{half.range}{If \code{TRUE} and a single population is found, the
normal distribution will be fit using the left half of the population}

\item{sd}{In the case of a single population, this is the multiplier
for the standard deviation to estimate the upper edge of the
negative population, default value of 2.5}

\item{positive}{If \code{TRUE}, gate the positive population}

\item{method}{Either \code{"minimum"} or \code{"left"} to choose the break as the
minimum between two peaks or to treat the left-most population
as a Gaussian population of the negative population}

\item{plot}{If \code{TRUE}, show a plot of the data and breakpoint}

\item{legend}{If \code{TRUE}, include a legend in the plot (if \code{plot == TRUE})}

\item{filterId}{The name assigned to the filter (\code{character} value)}

\item{...}{Other arguments to the original function are accepted
such as \code{borderQuant}, \code{absolute}, and \code{refLine} but are discarded}
}
\value{
An object of class \code{\link[flowCore:rectangleGate]{flowCore::rectangleGate}}
}
\description{
This function replaces \code{\link[flowStats:rangeGate]{flowStats::rangeGate()}} which never seemed to
work for me. It does not perform all the functions of the original
but will return a mores sensible \code{rectangleGate} object more
often than \code{\link[flowStats:rangeGate]{flowStats::rangeGate()}}.
}
\details{
This is a replacement function for \link[flowStats:rangeGate]{flowStats::rangeGate} that may be
more robust than the original function.

This function seeks the most likely breakpoint between two
populations to differentiate between the positive and negative
population. The
"negative" population is assumed to have a lower values and would be
located on the left of a histogram.

The default method (\code{method = "minimum"}) attempts to identify the
lowest minimum between the regions of greatest density or peaks on
a densityplot. The search can be limited to values between \code{from} and
\code{to}. If more than four high density regions exist \emph{or} if only a
single high density region is detected, the breakpoint will be
determined from the properties of the leftmost population where this
population is assumed to follows a normal distribution. This approach
also can be specified directly by \code{method = "left"}. The maximum of
this population is determined from a kernel density estimate. If
\code{half.range = TRUE}, the \emph{left} half of the population will be fit to
the presumed Gaussian distribution to estimate the standard deviation
of the population. With \code{method = "left"}, the value returned is the
position of the peak + \code{sd} times the standard deviation of the
distribution.

A diagnostic plot will be generated with base graphics if \code{plot = TRUE}.
If a plot is produced, the arguments \code{legend = TRUE} (default) will add
an informative legend to the plot showing the data and breakpoint. The
plotting option can be useful to iteratively adjust the search parameters
such as \code{adjust}, which is handed to the kernel density function,
of \code{sd}, \code{from} and \code{to}.
}
\examples{

}
\seealso{
\code{\link[flowStats:rangeGate]{flowStats::rangeGate()}}
}
