% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSet.R
\name{readSet}
\alias{readSet}
\title{Read and Transform Flow Set}
\usage{
readSet(path = ".", pattern = "\\\\d\\\\d\\\\d$", log.chan = NULL,
  alter.names = TRUE, name.keyword = "SAMPLE ID",
  phenoData = list(date = "$DATE", sample = "SAMPLE ID"),
  tfun = c("asinh", "log10"), depth = 10, log.cutoff = 2^depth, ...)
}
\arguments{
\item{path}{A character vector identifying the path to search for the FSC
files, defaults to current directory}

\item{pattern}{Grep pattern to identify fsc files, defaults to
\code{"\\d\\d\\d$"}}

\item{log.chan}{Character vector of channels to be log-transformed (after
"alter.names")}

\item{alter.names}{Logical value passed to \code{read.flowSet} to change "-"
to "." in names, default value of \code{TRUE}}

\item{name.keyword}{Character string for the  FCS keyword used for names,
default "SAMPLE ID"}

\item{phenoData}{List to serve as the rudiments of the phenoData built from 
keywords}

\item{tfun}{Transformation function to apply to log-transformed values,
either \ code{asinh} (default) or \code{log10}}

\item{depth}{Integer indicating bit depth (2^depth), 10 for FACVSCaliber, 24
for Acurri}

\item{log.cutoff}{Values greater than this value force log-transformation if 
the argument \code{log.chan} is \code{NULL}, defaults to \code{2^depth}}

\item{...}{Additional arguments to \code{\link{read.Flowset}}}
}
\value{
Flow Set with transformed and name-adjusted parameters.
}
\description{
Read a flow set and transform values with maximum values above given cutoff.
}
